requirements.txt
glob2
torch>=1.7.1
facenet-pytorch
Pillow
torchvision
